pyproject.toml
[project]
name = "llm-ltl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core scientific computing
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "numba>=0.57.0",
    # MuJoCo simulation (new open-source version, no license needed)
    "mujoco>=3.0.0",
    "glfw>=2.6.0",
    # Reinforcement learning
    "gymnasium>=0.29.0",
    # Deep learning (CUDA 12.x support for Blackwell architecture)
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    # Utilities
    "cloudpickle>=3.0.0",
    "joblib>=1.3.0",
    "h5py>=3.10.0",
    "matplotlib>=3.8.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.66.0",
    # Jupyter support
    "jupyter>=1.0.0",
    "ipython>=8.0.0",
    # Visualization
    "tensorboard>=2.15.0",
    # Local packages
    "maple",
    "robosuite",
    "gtimer>=1.0.0b5",
    "imageio[ffmpeg]>=2.37.2",
    "sk-video>=1.1.10",
    "pandas>=2.3.3",
]

[tool.uv.sources]
maple = { path = "maple", editable = true }
robosuite = { path = "robosuite", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["llm_ltl", "mujoco_py"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
